refactor: 비교 로직 공통화
